+++ src/FormController.js
@@ -90,7 +90,7 @@ export class FormController {
     // 6. It wipes out sync error
     this.validationRequests = new Map();
 
-    this.infoRequests = new Map();
+    this.dataRequests = new Map();
 
     // For array fields lol
     this.removalLocked = undefined;
@@ -111,7 +111,7 @@ export class FormController {
       maskedValues: {},
       dirt: {},
       focused: {},
-      info: {},
+      data: {},
       initialValues: this.options.current.initialValues || {}
     };
 
@@ -156,15 +156,15 @@ export class FormController {
     this.touchAllFields = this.touchAllFields.bind(this);
     this.keyDown = this.keyDown.bind(this);
     this.validateAsync = this.validateAsync.bind(this);
-    this.gatherInfo = this.gatherInfo.bind(this);
+    this.gatherData = this.gatherData.bind(this);
     this.validated = this.validated.bind(this);
     this.debouncedValidateAsync = debounceByName(this.validateAsync);
-    this.debouncedGatherInfo = debounceByName(this.gatherInfo);
+    this.debouncedGatherInfo = debounceByName(this.gatherData);
     this.getOptions = this.getOptions.bind(this);
     this.validateField = this.validateField.bind(this);
     this.getErrorMessage = this.getErrorMessage.bind(this);
     this.clearValue = this.clearValue.bind(this);
-    this.getInfo = this.getInfo.bind(this);
+    this.getData = this.getData.bind(this);
   }
 
   getOptions() {
@@ -325,7 +325,7 @@ export class FormController {
       meta.onChange(fieldState, e);
     }
 
-    if (meta.gatherInfo) {
+    if (meta.gatherData) {
       // Get error to determine if we even want to validateAsync
       this.debouncedGatherInfo(name);
     }
@@ -466,8 +466,8 @@ export class FormController {
     this.emit('field', name);
   }
 
-  getInfo(name) {
-    return ObjectMap.get(this.state.info, name);
+  getData(name) {
+    return ObjectMap.get(this.state.data, name);
   }
 
   getError(name) {
@@ -518,7 +518,7 @@ export class FormController {
       setError: this.setError,
       getFocused: this.getFocused,
       setFocused: this.setFocused,
-      getInfo: this.getInfo,
+      getData: this.getData,
       resetField: this.resetField,
       reset: this.reset,
       getFormState: this.getFormState,
@@ -547,7 +547,7 @@ export class FormController {
     const touched = !!this.getTouched(name);
     const pristine = !dirty;
     const validating = !!this.validationRequests.get(name);
-    const gathering = !!this.infoRequests.get(name);
+    const gathering = !!this.dataRequests.get(name);
 
     let showError = false;
     if (meta && meta.showErrorIfError) {
@@ -566,7 +566,7 @@ export class FormController {
       maskedValue: this.getMaskedValue(name),
       touched,
       error: this.getError(name),
-      info: this.getInfo(name),
+      data: this.getData(name),
       pristine,
       dirty,
       valid,
@@ -595,7 +595,7 @@ export class FormController {
       debug('Delete Focused', name);
       ObjectMap.delete(this.state.focused, name);
       debug('Delete Info', name);
-      ObjectMap.delete(this.state.info, name);
+      ObjectMap.delete(this.state.data, name);
 
       // Remember to update valid
       this.state.valid = ObjectMap.empty(this.state.errors);
@@ -619,7 +619,7 @@ export class FormController {
     ObjectMap.swap(this.state.errors, name, a, b);
     ObjectMap.swap(this.state.dirt, name, a, b);
     ObjectMap.swap(this.state.focused, name, a, b);
-    ObjectMap.swap(this.state.info, name, a, b);
+    ObjectMap.swap(this.state.data, name, a, b);
     // DO NOT emit event here we want to delay it on purpose because otherwise relevance will trigger with bad state
     // this.emit("field", name);
     this.state.pristine = false;
@@ -634,7 +634,7 @@ export class FormController {
     ObjectMap.delete(this.state.errors, name);
     ObjectMap.delete(this.state.dirt, name);
     ObjectMap.delete(this.state.focused, name);
-    ObjectMap.delete(this.state.info, name);
+    ObjectMap.delete(this.state.data, name);
     // DO NOT emit event here we want to delay it on purpose because otherwise relevance will trigger with bad state
     // this.emit("field", name);
     this.state.pristine = false;
@@ -770,14 +770,14 @@ export class FormController {
 
   gathered(name, res) {
     debug(
-      `Setting ${name}'s info to ${res} with ${
+      `Setting ${name}'s data to ${res} with ${
         this.state.gathering
       } gatherers left`
     );
-    ObjectMap.set(this.state.info, name, res);
+    ObjectMap.set(this.state.data, name, res);
 
     // Clear out validating
-    this.infoRequests.delete(name);
+    this.dataRequests.delete(name);
 
     // Always update
     this.emit('field', name);
@@ -847,7 +847,7 @@ export class FormController {
     }
   }
 
-  gatherInfo(name) {
+  gatherData(name) {
     debug('EXECUTING INFO ASYNC', name);
     // Get meta for field
     const meta = this.fieldsMap.get(name)?.current;
@@ -855,27 +855,27 @@ export class FormController {
     // Get the value
     const value = this.getValue(name);
 
-    if (meta && meta.gatherInfo) {
+    if (meta && meta.gatherData) {
       this.state.gathering = this.state.gathering + 1;
       const uuid = uuidv4();
-      debug('INFO REQUEST', uuid);
-      this.infoRequests.set(name, { uuid, value });
+      debug('DATA REQUEST', uuid);
+      this.dataRequests.set(name, { uuid, value });
 
       // Because we may have been debounced need to update field here
       this.emit('field', name);
 
       meta
-        .gatherInfo(value, this.state)
+        .gatherData(value, this.state)
         .then(res => {
           this.state.gathering = this.state.gathering - 1;
-          const stale = this.infoRequests.get(name).uuid !== uuid;
+          const stale = this.dataRequests.get(name).uuid !== uuid;
 
           // What in the hell is invalid and why do I need it??
           // because the value can be outdated
           const invalid =
-            this.infoRequests.get(name).value !== this.getValue(name);
+            this.dataRequests.get(name).value !== this.getValue(name);
           if (!stale && !invalid) {
-            debug('INFO FINISH', uuid);
+            debug('DATA FINISH', uuid);
             this.gathered(name, res);
           } else {
             debug(
@@ -888,11 +888,11 @@ export class FormController {
         })
         .catch(err => {
           this.state.gathering = this.state.gathering - 1;
-          const stale = this.infoRequests.get(name).uuid !== uuid;
+          const stale = this.dataRequests.get(name).uuid !== uuid;
           const invalid =
-            this.infoRequests.get(name).value !== this.getValue(name);
+            this.dataRequests.get(name).value !== this.getValue(name);
           if (!stale && !invalid) {
-            debug('INFO FINISH', uuid);
+            debug('DATA FINISH', uuid);
             this.gathered(name, err.message);
           } else {
             debug(
@@ -922,7 +922,7 @@ export class FormController {
       maskedValues: {},
       dirt: {},
       focused: {},
-      info: {},
+      data: {},
       initialValues: this.options.current.initialValues || {}
     };